I18N is enabled in the demo app, in a rough way. A French translation (almost complete) of the UI is included. There is also an action to switch from the default (en) to fr:
http://demo/culture/switch?culture=fr

Note that (as always) Apostrophe also i18n's slot content, so when you first switch you'll be looking at empty pages. That's by design, edit them to create the French version of those pages, etc. This feature has been in Apostrophe for quite a while, it's the French UI that is new.